feat: make lazy imports as default imports
