--- src/pages/CvProfile/CvProfile.tsx
@@ -1,12 +1,0 @@
-import {FC} from "react";
-import {Outlet} from "react-router-dom";
-import {CvsTabs} from "@/pages/Cvs/CvsTabs";
-
-export const CvProfile: FC = () => {
-  return (
-    <>
-      <CvsTabs />
-      <Outlet />
-    </>
-  );
-};
