--- src/pages/Profile/Profile.tsx
@@ -1,12 +1,0 @@
-import {FC} from "react";
-import {Outlet} from "react-router-dom";
-import {ProfileTabs} from "@/pages/Profile/ProfileTabs/ProfileTabs";
-
-export const Profile: FC = () => {
-  return (
-    <>
-      <ProfileTabs />
-      <Outlet />
-    </>
-  );
-};
