+++ src/pages/Profile/index.tsx
@@ -1,3 +1,12 @@
-import {Profile} from "./Profile";
+import {FC} from "react";
+import {Outlet} from "react-router-dom";
+import {ProfileTabs} from "@/pages/Profile/ProfileTabs/ProfileTabs";
 
-export default Profile;
+export const Profile: FC = () => {
+  return (
+    <>
+      <ProfileTabs />
+      <Outlet />
+    </>
+  );
+};
